feat(gitpython): introduce the python library.
[GitPython](https://gitpython.readthedocs.io) is a python library used to
interact with git repositories, high-level like git-porcelain, or low-level like
git-plumbing.

It provides abstractions of git objects for easy access of repository data, and
additionally allows you to access the git repository more directly using either
a pure python implementation, or the faster, but more resource intensive git
command implementation.

Explain how to:

* Initialize or load repositories.
* Make commits.
* Interact with the history.
* Test applications that use it.

style(pytest): add freezegun link.

feat(python_snippets): add today's learned python tricks

* [Get system's timezone and use it in
datetime](python_snippets.md#get-systems-timezone-and-use-it-in-datetime).
* [Capitalize a sentence](python_snippets.md#capitalize-a-sentence).
* [Get the last monday
datetime](python_snippets.md#get-the-last-monday-datetime).

feat(pilates): introduce the art.

[Pilates](https://en.wikipedia.org/wiki/Pilates) is a physical fitness system
based on controlled movements putting emphasis on alignment, breathing,
developing a strong core, and improving coordination and balance. The core (or
powerhouse), consisting of the muscles of the abdomen, low back, and hips, is
thought to be the key to a person's stability.

Pilates' system allows for different exercises to be modified in range of
difficulty from beginner to advanced or to any other level, and also in terms of
the instructor and practitioner's specific goals and/or limitations. Intensity
can be increased over time as the body adapts itself to the exercises.

You can think of yoga, but without the spiritual aspects.

Also added:

* It's principles
* The [swing from table](pilates.md#swing-from-table) exercise.

feat(adr#adr-template): update the ADR template with the week learnings.

* Add the Proposals and Date sections
* Explain the possible Status states.
* Add an [Ultisnip](https://github.com/SirVer/ultisnips) vim snippet.
* Explain how I've used it to create [mkdocs-newsletter](https://github.com/lyz-code/mkdocs-newsletter).