feat(monitoring_comparison): Compare Nagios and Prometheus as monitoring tools
TL;DR: Prometheus is better, but it needs more effort.

Nagios is suitable for basic monitoring of small and/or static systems where
blackbox probing is sufficient.

If you want to do whitebox monitoring, or have a dynamic or cloud based
environment, then Prometheus is a good choice.

fix(adr): update the template and add some tips.

fix(gitpython): improve the way to test gitpython programs

Add the test data on each tests instead of the monolith fixture proposed
in the first version.

fix(type_hints): remove the warning of pydantic not supporting TypedDicts

The issues are solved and the PR are merged, so it's now supported.

feat(kubernetes): Add some tools to test.

* [Popeye](https://github.com/derailed/popeye) is a utility that scans live
    Kubernetes cluster and reports potential issues with deployed resources and
    configurations.

* [Stern](https://github.com/wercker/stern) allows you to tail multiple pods on
    Kubernetes and multiple containers within the pod. Each result is color
    coded for quicker debugging.

* [Fairwinds' Polaris](https://github.com/FairwindsOps/polaris) runs
    a variety of checks to ensure that Kubernetes pods and controllers are
    configured using best practices, helping you avoid problems in the
    future.

* [kube-hunter](https://github.com/aquasecurity/kube-hunter) hunts for security
    weaknesses in Kubernetes clusters. The tool was developed to increase
    awareness and visibility for security issues in Kubernetes environments.

feat(gadgetbridge): introduce the android band application and how to use it.

[Gadgetbridge](https://gadgetbridge.org/) is an Android (4.4+) application which
will allow you to use your Pebble, Mi Band, Amazfit Bip and HPlus device (and
more) without the vendor's closed source application and without the need to
create an account and transmit any of your data to the vendor's servers.

feat(issues): monitor the week new issues

Regarding:

* Gitea Kanban board improvements
* Docker monitorization
* GadgetBridge improvements

feat(mkdocs): add developer information on how to build plugins

Analyze the Files, File objects and the on_files event.

feat(virtual_assistant): add mycroft to the comparison of solutions