feat(pytest#Running tests in parallel): Explain how to run tests in parallel
`pytest-xdist` makes it possible to run the tests in parallel, useful when the
test suit is large or when the tests are slow.

```bash
pip install pytest-xdist
pytest -n auto
```

fix(boto3#test-route53): Add note that pagination is not yet supported when testing route53

I've [opened an issue](https://github.com/spulec/moto/issues/3879) to
solve it.