feat(orthography#stop-saying-i-know): explain what can you use instead of I know
Using "I know" may not be the best way to show the other person that you've got
the information. You can take the chance to use other words that additionally
gives more context on how you stand with the information you've received,
thus improving the communication and creating a bond.