feat(projects#self-hosted-map): introduce seedling self-hosted map project
I love maps, as well as traveling and hiking. This project aims to create a web
interface that let's me interact with the data gathered throughout my life. I'd
like to:

* Browse the waypoints and routes that I've done.
* Create routes and export the gpx.
* Be able to search through the data
* Plan trips

feat(projects#switch-to-a-better-browser): Introduce the seed project to
switch to a better browser

Firefox + Trydactil has it's limits, and Firefox has been following an ill path
for a while, [qutebrowser](https://github.com/qutebrowser/qutebrowser) looks
like the perfect replacement.