feat(pytest#Setting a timeout for your tests): Explain how to set a timeout for your tests
Using [pytest-timeout](https://pypi.org/project/pytest-timeout/).

feat(pytest#Rerun tests that fail sometimes): explain how to rerun tests that fail sometimes.

With [pytest-rerunfailures](https://pypi.org/project/pytest-rerunfailures/) that is
a plugin for pytest that re-runs tests to eliminate intermittent failures. Using
this plugin is generally a bad idea, it would be best to solve the reason why
your code is not reliable. It's useful when you rely on non robust third party
software in a way that you can't solve, or if the error is not in your code but
in the testing code, and again you are not able to fix it.

feat(python_snippets#Create combination of elements in groups of two):
Explain how to create combination of elements in groups of two

```python
>>> list(itertools.combinations('ABC', 2))
[('A', 'B'), ('A', 'C'), ('B', 'C')]
```

feat(selenium#Set timeout of a response): explain how to Set timeout of a response

```python
driver.set_page_load_timeout(30)
```

feat(selenium#Chromedriver hangs up unexpectedly): explain how to fix when Chromedriver hangs up unexpectedly

```python
os.environ["DBUS_SESSION_BUS_ADDRESS"] = "/dev/null"
```