feat(antitransphobia): Introduce Anti-transphobia
Anti-transphobia being reductionist is the opposition to the collection of ideas
and phenomena that encompass a range of negative attitudes, feelings or actions
towards transgender people or transness in general. Transphobia can include
fear, aversion, hatred, violence, anger, or discomfort felt or expressed towards
people who do not conform to social gender expectations. It is often expressed
alongside homophobic views and hence is often considered an aspect of
homophobia.

feat(antitransphobia#terf): Introduce arguments against terf ideology

[TERF](https://en.wikipedia.org/wiki/TERF) is an acronym for *trans-exclusionary
radical feminist*. The term originally applied to the minority of feminists that
expressed transphobic sentiments such as the rejection of the assertion that
trans women are women, the exclusion of trans women from women's spaces, and
opposition to transgender rights legislation. The meaning has since expanded to
refer more broadly to people with trans-exclusionary views who may have no
involvement with radical feminism.