feat(antifascist_actions#A fake company and five million recycled flyers): A fake company and five million recycled flyers
A group of artists belonging to the [Center for political
beauty](https://politicalbeauty.com/) created a fake company [Flyerservice
Hahn](https://www.flyerservice-hahn.de/) and convinced more than 80 regional
sections of the far right party AfD to hire them to deliver their electoral
propaganda.

They gathered five million flyers, with a total weight of 72 tons. They justify
that they wouldn't be able to lie to the people, so they did nothing in the
broader sense of the word. They declared that they are the ["world wide leader in
the non-delivery of nazi
propaganda"](https://www.youtube.com/watch?v=qcBVq-PtrpY). At the start of the
electoral campaign, they went to the AfD stands, and they let their members to
give them flyers the throw them to the closest bin. "It's something that any
citizen can freely do, we have only industrialized the process".

They've done a [crowdfunding](https://afd-muell.de/) to fund the legal process
that may result.