feat(ansible_snippets#Speed up the stat module): Speed up the stat module
The `stat` module calculates the checksum and the md5 of the file in order to
get the required data. If you just want to check if the file exists use:

```yaml
- name: Verify swapfile status
  stat:
    path: "{{ common_swapfile_location }}"
    get_checksum: no
    get_md5: no
    get_mime: no
    get_attributes: no
  register: swap_status
  changed_when: not swap_status.stat.exists
```

feat(pdm): Introduce PDM

[PDM](https://pdm.fming.dev/) is a modern Python package manager with [PEP
582](https://www.python.org/dev/peps/pep-0582/) support. It installs and manages
packages in a similar way to npm that doesn't need to create a virtualenv at
all!

feat(vim#Delete a file inside vim): Delete a file inside vim

```vim
:call delete(expand('%')) | bdelete!
```