feat(projects#life-warnings): Create the Life Warnings seed
I've always tackled the pursuit of the peace of mind by improving in task
management, for example trying to predict when I have to do something in order
to avoid a nuisance. Maybe it's more interesting to monitor and visibilice the
warnings that are affecting you.

refactor(projects): Merge Self host a routing web application seed with host my own map seedling