feat(code_learning): Introduce guidelines to learn how to code
Learning to code is a never ending, rewarding, frustrating, enlightening task.
In this article you can see what is the generic roadmap (in my personal opinion)
of a developer. As each of us is different, probably a generic roadmap won't
suit your needs perfectly, if you are new to coding, I suggest you find
a [mentor](mentoring.md) so you can both tweak it to your case.

feat(frontend_learning): Introduce guidelines to learn how to become a frontend developer

This section is the particularization of the [Development learning article](code_learning.md) for a frontend developer, in particular a [Vue](vuejs.md) developer.

A Front-End Developer is someone who creates websites and web applications. It's
main responsibility is to create what the user sees.

The basic languages for Front-End Development are [HTML](html.md),
[CSS](css.md), and [JavaScript](javascript.md). Nowadays writing interfaces with
only the basic languages makes no sense as there are other languages and
frameworks that make better and quicker solutions. One of them is
[Vue](vuejs.md), which is the one I learnt, so the whole document will be
focused on this path, nevertheless there are others popular ones like:
[Bootstrap](https://www.w3schools.com/whatis/whatis_bootstrap.asp),
[React](react.md), [jQuery](https://www.w3schools.com/jquery/default.asp) or
[Angular](https://www.w3schools.com/jquery/default.asp).

The difference between Front-End and Back-End is that Front-End refers to how
a web page looks, while back-end refers to how it works.

feat(mentoring): Introduce the concept and guidelines of mentorship

[Mentoring](https://en.wikipedia.org/wiki/Mentorship) is a process for the
informal transmission of knowledge, social capital, and the psychosocial support
perceived by the recipient as relevant to work, career, or professional
development; mentoring entails informal communication, usually face-to-face and
during a sustained period of time, between a person who is perceived to have
greater relevant knowledge, wisdom, or experience (the mentor) and a person who
is perceived to have less (the apprentice).

feat(park_programming): Introduce the park programming concept

Park programming is as you may guess, programming in parks. It includes
guidelines on:

* How to park program
* Finding the best spots
* Finding the best times