fix(helmfile): tweak the update process to make it more resilient
* Check that all the helm deployments are well deployed with `helm list -A | grep -v deployed`
* Change the context to the production cluster before running the production steps

feat(helmfile#Cannot patch X field is immutable): Fix Cannot patch X field is immutable error

You may think that deleting the resource, usually a deployment or daemonset will
fix it, but `helmfile apply` will end without any error, the resource won't be recreated
, and if you do a `helm list`, the deployment will be marked as failed.

The solution we've found is disabling the resource in the chart's values so that
it's uninstalled an install it again.

This can be a problem with the resources that have persistence. To patch it,
edit the volume resource with `kubectl edit pv -n namespace volume_pvc`, change
the `persistentVolumeReclaimPolicy` to `Retain`, apply the changes to uninstall,
and when reinstalling configure the chart to use that volume (easier said than
done).

feat(cruft): use `skip` to avoid problems with `.git`

Since `2.10.0` you can add a `skip` category inside the `.cruft.json`, so that it doesn't check that
directory:

```json
{
  "template": "xxx",
  "commit": "xxx",
  "checkout": null,
  "context": {
    "cookiecutter": {
       ...
    }
  },
  "directory": null,
  "skip": [
    ".git"
  ]
}
```

feat(python): add huey

[huey](https://github.com/coleifer/huey) is a little task queue for python.

feat(use_warnings): Change signature of method if you can

You can take the chance of the deprecation to change the signature of
the function, so that if the user is using the old argument, it uses the
old behaviour and gets the warning, and if it's using the new argument,
it uses the new. The advantage of changing the signature is that you
don't need to do another deprecation for the temporal argument flag.