feat(calendar_management): Introduce Calendar Management
Since the break of my taskwarrior instance I've used a physical calendar to
manage the tasks that have a specific date. Can't wait for
the first version of [`pydo`](https://github.com/lyz-code/pydo) to be finished.

The next factors made me search for a temporal solution:

* It's taking longer than expected.
* I've started using a nextcloud calendar with some friends.
* I frequently use Google calendar at work.
* I'm sick of having to log in Nexcloud and Google to get the day's
    appointments.

To fulfill my needs the solution needs to:

* Import calendar events from different sources, basically through
    the [CalDAV](http://en.wikipedia.org/wiki/CalDAV) protocol.
* Have a usable terminal user interface
* Optionally have a command line interface or python library so it's easy to make scripts.
* Optionally it can be based in python so it's easy to contribute
* Support having a personal calendar mixed with the shared ones.
* Show all calendars in the same interface

Looking at the available programs I found [`khal`](khal.md), which looks like
it may be up to the task.

Go through the [installation](khal.md#installation) steps and configure the
instance to have a local calendar.

If you want to sync your calendar events through CalDAV, you need to set
[vdirsyncer](vdirsyncer.md).

feat(khal): Introduce khal

[`khal`](https://khal.readthedocs.io/en/latest/index.html) is a standards based
Python CLI (console) calendar program, able to synchronize with [CalDAV](http://en.wikipedia.org/wiki/CalDAV) servers through
[`vdirsyncer`](vdirsyncer.md).

Features:

* Can read and write events/icalendars to vdir, so [`vdirsyncer`](vdirsyncer.md)
    can be used to synchronize calendars with a variety of other programs, for
    example CalDAV servers.
* Fast and easy way to add new events
* `ikhal` (interactive `khal`) lets you browse and edit calendars and events.

Limitations:

* Only rudimentary support for creating and editing recursion rules
* You cannot edit the timezones of events

feat(vdirsyncer): Introduce vdirsyncer

[vdirsyncer](https://vdirsyncer.pimutils.org/en/stable/) is a Python
command-line tool for synchronizing calendars and addressbooks between a variety
of servers and the local filesystem. The most popular usecase is to synchronize
a server with a local folder and use a set of other programs such as
[`khal`](khal.md) to change the local events and contacts. Vdirsyncer can then
synchronize those changes back to the server.

However, `vdirsyncer` is not limited to synchronizing between clients and
servers. It can also be used to synchronize calendars and/or addressbooks
between two servers directly.

It aims to be for calendars and contacts what OfflineIMAP is for emails.