feat(grapheneos): Introduce GrapheneOS
[GrapheneOS](https://grapheneos.org/) is a private and secure mobile operating
system with Android app compatibility. Developed as a non-profit open source
project.

feat(mdformat): Introduce MDFormat

[MDFormat](https://mdformat.readthedocs.io/en/stable/) is an opinionated
Markdown formatter that can be used to enforce a consistent style in Markdown
files. Mdformat is a Unix-style command-line tool as well as a Python library.

The features/opinions of the formatter include:

* Consistent indentation and whitespace across the board
* Always use ATX style headings
* Move all link references to the bottom of the document (sorted by label)
* Reformat indented code blocks as fenced code blocks
* Use 1. as the ordered list marker if possible, also for noninitial list items.

It's based on the
[`markdown-it-py`](https://markdown-it-py.readthedocs.io/en/latest/index.html)
Markdown parser, which is a Python implementation of
[`markdown-it`](https://github.com/markdown-it/markdown-it).

feat(redox): Introduce Redox

[Redox](https://github.com/mattdibi/redox-keyboard) is an awesome Do It Yourself
mechanical keyboard