feat(botany#Alder): Introduce the Alder tree
![ ](Alder-3.jpg)

Alders are trees comprising the genus Alnus in the birch family Betulaceae (like
the [birch](#birch)). The genus parts are "al" which means "close by" and "lan"
which means "side of the river", so they are trees that grow close to rivers or
creeks.

fix(botany#beech): Give more details of the beech tree

![ ](beech-4.jpg)
The leaves of beech trees are elliptic, a little pointy at the end, flat, and
with a short petiole. They are big and wide leaves ranging from 4-9 cm long.
Very abundant, they have a light green colour with a darker tone and glossy on
the upper side.

The fruit is a small, sharply three-angled nut 10-15 mm long, borne singly or in
pairs in soft-spined husks 1.5-2.5 cm long, known as cupules. The husk can have
a variety of spine- to scale-like appendages, the character of which is, in
addition to leaf shape, one of the primary ways beeches are differentiated. The
nuts are edible, though bitter (though not nearly as bitter as acorns) with a
high tannin content, and are called beechnuts or beechmast.

They are big trees easily going between 30 and 45 meters. It looks very
different if its isolated or being part of a forest. The first one the branches
grow from the middle of the trunk and are horizontal, in the second, the
branches go up and start over the half of the trunk. The principal root is very
powerful, with very strong secondary roots, developing lateral superficial
roots.

![ ](beech-bark.jpg)

The trunk is right with a grayish-cinder bark, smooth until it's old, usually
covered by moss an lichen. Smaller branches are zigzagging with reddish-brown
pointy buds.

The canopy is big, dense, rounded and semi spheric, giving a lot of shadow.

It grows slow in the first years, being the most active between the tenth and
twelve year, reaching it's maximum height when it's twenty five, although it
lives around three hundred years.

fix(botany#birch): Give more details of the birch tree

The simple leaves are rhomboidal, between 3 and 6 cm, singly or doubly serrate
except at the base, feather-veined, petiolate and stipulate. Although they are
alternate, many leaves spawn from each side of the branch, making some think
that they are not alternate. They often appear in pairs, but these pairs are
really borne on spur-like, two-leaved, lateral branchlets.

The canopy is rounded and irregular giving few shadow.

The fruit is a small samara, although the wings may be obscure in some species.
They differ from the alders in that the female catkins are not woody and
disintegrate at maturity, falling apart to release the seeds, unlike the woody,
cone-like female alder catkins.

The bark of all birches is characteristically smooth and white, although in
older ones the lower part is usually cracked and takes blackish brown colours.
It's marked with long, horizontal lenticels, and often separates into thin,
papery plates, especially upon the paper birch.

feat(botany#How to tell apart the different trees): How to tell apart the different trees

Alder vs Beech:

| Property                      | Beech                     | Alder                     |
| ----------------------------- | ------------------------- | ------------------------- |
| Leaf border                   | flat                      | sparsely toothed          |
| Leaf form                     | elliptic                  | rounded                   |
| Same colour both sides        | no (darker and glossy up) | yes                       |
| Sticky leafs                  | no                        | yes                       |
| Size                          | 30-45m                    | 10-12m (in Europe)        |
| Knots on the roots with fungi | no                        | yes                       |
| Where they grow               | everywhere                | close to rivers or creeks |

Alder vs Birch:

| Property        | Birch                    | Alder                     |
| --------------- | ------------------------ | ------------------------- |
| Leaf border     | heavy toothed            | sparsely toothed          |
| Leaf form       | rhomboidal               | rounded                   |
| Sticky leafs    | no                       | yes                       |
| Where they grow | very close to each other | close to rivers or creeks |

Beech vs Birch:

| Property               | Beech                     | Birch                    |
| ---------------------- | ------------------------- | ------------------------ |
| Leaf border            | flat                      | heavy toothed            |
| Leaf form              | elliptic                  | rhomboidal               |
| Size                   | 30-45m                    | 10-15m (in Europe)       |
| Same colour both sides | no (darker and glossy up) | yes                      |
| Where they grow        | everywhere                | very close to each other |

feat(linux_snippets#Download TS streams): Download TS streams

Some sites give stream content with small `.ts` files that you can't download
directly. Instead open the developer tools, reload the page and search for
a request with extension `.m3u8`, that gives you the playlist of all the chunks
of `.ts` files. Once you have that url you can use `yt-dlp` to download it.